// Cxh_pulse_chirstmas_edition/frontend/src/types/sections.types.ts
export type IndicatorTrendsTableProps = {
  dateRange: { from: Date; to: Date };
  ward?: string;
  county?: string;
  selectedIndicator?: string; // Filter by selected indicator, if empty/null show all
  hideTitle?: boolean;
};